requirements.txt
# Odyssey ML SDK
git+https://github.com/odysseyml/odyssey-python.git

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Video Processing (Optional)
opencv-python>=4.8.0

# Async utilities
aiohttp>=3.9.0
aiofiles>=23.2.0

# Environment management
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0
